# service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ",
    "version": "0.1.0",
    "description": "Сверка проектной, рабочей и исполнительной документации. Асинхронная pull-модель: загрузка возвращает process_id, клиент опрашивает статус и забирает протокол."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "ProcessStatus": {
        "type": "object",
        "required": [
          "process_id",
          "object_id",
          "status",
          "upload_status",
          "processing",
          "files",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "process_id": {
            "type": "string",
            "format": "uuid"
          },
          "object_id": {
            "type": "string"
          },
          "object_name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PARSING",
              "READY",
              "VERIFYING",
              "COMPLETED",
              "FINALIZED"
            ]
          },
          "scenario": {
            "type": "string",
            "nullable": true
          },
          "upload_status": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reading_mode": {
            "type": "string",
            "nullable": true,
            "description": "Способ чтения объекта, выбранный при загрузке (#54), одно из layer, tesseract, model: layer — только текстовый слой, tesseract — слой и распознавание сканов, model — слой, распознавание и модель. null — читается способом сервиса"
          },
          "model_name": {
            "type": "string",
            "nullable": true,
            "description": "Модель, выбранная при загрузке для режима «модель» (#54); null — модель сервиса. Чем объект прочитан на самом деле, видно в readiness.model_name после разбора"
          },
          "processing": {
            "$ref": "#/components/schemas/Processing"
          },
          "progress": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "properties": {
              "step": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "done": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "updated_at": {
                "type": "string"
              }
            }
          },
          "files": {
            "type": "object",
            "required": [
              "accepted",
              "rejected",
              "bytes"
            ],
            "properties": {
              "accepted": {
                "type": "integer"
              },
              "rejected": {
                "type": "integer"
              },
              "bytes": {
                "type": "integer"
              }
            }
          },
          "protocol_version": {
            "type": "integer",
            "nullable": true
          },
          "protocol": {
            "type": "object",
            "description": "Последняя версия протокола и версии Матрицы, набора данных и конвейера, по которым она собрана",
            "properties": {
              "version": {
                "type": "integer"
              },
              "status": {
                "type": "string"
              },
              "matrix_version": {
                "type": "string",
                "nullable": true
              },
              "dataset_version": {
                "type": "string",
                "nullable": true
              },
              "model_version": {
                "type": "string",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "findings": {
            "$ref": "#/components/schemas/FindingCounts"
          },
          "coverage": {
            "type": "object",
            "nullable": true,
            "description": "Охват Матрицы последней версией протокола: compared — параметров, по которым значения сопоставлены (правило есть и хотя бы одна запись сравнена), total — параметров в Матрице. Считается так же, как группа COMPARED дашборда и раздела 2 протокола. Без сравнения — параметры с правилом прода, по которым значения не сопоставлены: missing_evidence — нет доказательств (значения или документа нет в одной из стадий), not_comparable — значения несопоставимы; как одноимённые группы дашборда.",
            "properties": {
              "compared": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "missing_evidence": {
                "type": "integer"
              },
              "not_comparable": {
                "type": "integer"
              }
            }
          },
          "readiness": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Готовность объекта (#39): pages, pages_without_text, share_without_text, stage_unknown, section_other, unsupported, pages_low_quality, ocr_enabled, model_enabled, by_text_source, reading_mode и model_name — чем объект прочитан на самом деле (#54). Считается при разборе, лежит в отчёте протокола"
          },
          "recompute": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Что пересчитано при последней обработке, а что перенесено из прошлой версии (#60): full, reason, documents_read, documents_total, recomputed, carried_over. Дозагрузка пересчитывает только параметры, которых коснулись новые документы"
          },
          "sync": {
            "type": "object",
            "nullable": true,
            "description": "Передача финализированного протокола в ИАИС «РиН» (ТЗ 9.6): PENDING_SYNC, SYNCED, REJECTED, CANCELLED",
            "properties": {
              "status": {
                "type": "string"
              },
              "protocol_version": {
                "type": "integer"
              },
              "attempts": {
                "type": "integer"
              },
              "max_retries": {
                "type": "integer"
              },
              "next_attempt_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "last_error": {
                "type": "string",
                "nullable": true
              },
              "last_status_code": {
                "type": "integer",
                "nullable": true
              },
              "external_id": {
                "type": "string",
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "synced_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "finalized_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/objects": {
      "get": {
        "summary": "Объекты для дашборда",
        "tags": [
          "objects"
        ],
        "description": "Объекты с последним процессом проверки: его статус и счётчики находок дают светофор.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "objects"
                  ],
                  "properties": {
                    "objects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "object_id",
                          "name",
                          "processes"
                        ],
                        "properties": {
                          "object_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "processes": {
                            "type": "integer"
                          },
                          "last_process": {
                            "$ref": "#/components/schemas/ProcessStatus"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "summary": "Уведомления своей роли",
        "tags": [
          "system"
        ],
        "description": "Инспектор получает уведомление о готовности протокола, администратор — о сбоях обработки после исчерпания повторов (ТЗ: таймаут — до двух повторов, затем уведомление администратора).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "notifications"
                  ],
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "role",
                          "level",
                          "message",
                          "created_at"
                        ],
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "role": {
                            "type": "string"
                          },
                          "level": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                            "description": "ACTION — требует действия, QUALITY — наблюдение о разборе (живёт в журнале), INFO"
                          },
                          "message": {
                            "type": "string"
                          },
                          "process_id": {
                            "type": "string",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "summary": "Журнал аудита процесса (администратор)",
        "tags": [
          "system"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "process_id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entries"
                  ],
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "user_id": {
                            "type": "string",
                            "nullable": true
                          },
                          "action": {
                            "type": "string"
                          },
                          "details": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "ip_address": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/objects/{id}": {
      "delete": {
        "summary": "Удалить черновой объект со всеми его проверками (администратор)",
        "tags": [
          "objects"
        ],
        "description": "Удаляются объекты, у которых нет финализированных протоколов: загрузили не тот пакет, перепутали идентификатор, пробный прогон. Финализированный протокол — документ с версией, решениями инспектора и записью в ИАИС «РиН», поэтому такой объект не удаляется (409 OBJECT_NOT_DELETABLE); идущая обработка тоже мешает. Вместе с объектом уходят его процессы, файлы, находки и протоколы; журнал аудита, уведомления и выпущенные версии набора решений остаются — это записи о том, что делали, и снимки. Содержимое файлов из хранилища удаляется только то, на которое больше никто не ссылается: хранилище адресовано по SHA-256, и один и тот же файл может принадлежать другому объекту.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "object_id",
                    "processes",
                    "files",
                    "blobs_removed"
                  ],
                  "properties": {
                    "object_id": {
                      "type": "string"
                    },
                    "processes": {
                      "type": "integer"
                    },
                    "files": {
                      "type": "integer"
                    },
                    "blobs_removed": {
                      "type": "integer",
                      "description": "сколько файлов удалено из хранилища как никому не нужные"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}